refactor: move handling project-update events to useUpdateEvents
fixes a problem with project rows showing only the modified project for the user how modified it when moving the timeline
